Fix compilation warnings
* configure.ac: Check for intptr_t
* comsat/action.c (need_crlf): Use intermediate cast to intptr_t.
* imap4d/imap4d.c (cb2_group, mu_get_user_groups): Likewise.
* libmailutils/stream/file_stream.c (fd_ioctl): Likewise.
* libmailutils/stream/mapfile_stream.c (_mapfile_ioctl): Likewise.
* libmailutils/stream/tcp.c (_tcp_ioctl): Likewise.
* maidag/maidag.c (cb2_group): Likewise.

* frm/common.c (puts_bidi): Fix the type of 3rd parameter to
fribidi_log2vis

* include/mailutils/body.h: Use MU_DEPRECATED instead of\
 __attribute__ ((deprecated)).
* include/mailutils/cfg.h: Likewise.
* include/mailutils/header.h: Likewise.
* include/mailutils/libargp.h: Likewise.
* include/mailutils/mailbox.h: Likewise.
* include/mailutils/mailer.h: Likewise.
* include/mailutils/registrar.h: Likewise.
* include/mailutils/message.h: Likewise.
(mu_message_set_stream): Remove deprecation marker.

+++ comsat/action.c
@@ -186,7 +186,7 @@ need_crlf (mu_stream_t str)
 
   if (mu_stream_ioctl (str, MU_IOCTL_TRANSPORT, MU_IOCTL_OP_GET, trans))
     return 1; /* suppose we do need it */
-  if (tcgetattr ((int)trans[0], &tbuf) == 0 &&
+  if (tcgetattr ((int) (intptr_t) trans[0], &tbuf) == 0 &&
       (tbuf.c_oflag & OPOST) && (tbuf.c_oflag & ONLCR))
     return 0;
   else
